Rename pmcmean to pmc_mean for consistency
